peniquitous.js: Get rid of test code
This was used to test out simulating a down arrow key press on the
Google search home page with the field selected and something typed in.
If the first entry in the suggestion drop-down was selected after the
timeout fired, that means it worked.

We no longer need this test code.